splitted wevbiew compontnts
