+++ frontend/.vscode/setting.json
@@ -1,8 +1,6 @@
 {
   "editor.defaultFormatter": "esbenp.prettier-vscode",
-  "editor.formatOnPaste": true,
   "editor.formatOnSave": true,
-  "editor.formatOnType": true,
   "editor.codeActionsOnSave": [
     "source.organizeImports",
     "source.fixAll.eslint"
